@@ -14,7 +14,7 @@ description: A learning platform application for both web and mobile
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -40,6 +40,7 @@ dependencies:
   cached_network_image:
   flutter_svg:
   getflutter: ^1.0.11
+  universal_html: ^1.1.4
 
 dev_dependencies:
   flutter_test:
