@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:zengage_learning_platform/constants/app_colors.dart';
+import 'package:zengage_learning_platform/extensions/hover_extensions.dart';
 import 'package:zengage_learning_platform/routes/route_generator.dart';
 
 class BlueBorderWhiteButton extends StatelessWidget {
@@ -29,26 +30,28 @@ class BlueBorderWhiteButton extends StatelessWidget {
     if (height == -1.0) {
       height = MediaQuery.of(context).size.height;
     }
-    return Container(
-      color: Colors.white,
-      child: OutlineButton(
-        padding: EdgeInsets.fromLTRB(
-            width * 0.1, height * 0.025, width * 0.1, height * 0.025),
-        child: Text(
-          buttonText,
-          style: TextStyle(
-              fontSize: (width > 700.0) ? 18.0 : 12.0,
-              color: textGreyColor,
-              fontWeight: FontWeight.bold),
-        ),
-        onPressed: () {
-          print("Contact Us pressed");
-          Navigator.of(context).pushNamed(route);
-        },
-        borderSide: BorderSide(
-          color: Colors.lightBlue, //Color of the border
-          style: BorderStyle.solid, //Style of the border
-          width: 4, //width of the border
+    return HandCursor(
+      child: Container(
+        color: Colors.white,
+        child: OutlineButton(
+          padding: EdgeInsets.fromLTRB(
+              width * 0.1, height * 0.025, width * 0.1, height * 0.025),
+          child: Text(
+            buttonText,
+            style: TextStyle(
+                fontSize: (width > 700.0) ? 18.0 : 12.0,
+                color: textGreyColor,
+                fontWeight: FontWeight.bold),
+          ),
+          onPressed: () {
+            print("Contact Us pressed");
+            Navigator.of(context).pushNamed(route);
+          },
+          borderSide: BorderSide(
+            color: Colors.lightBlue, //Color of the border
+            style: BorderStyle.solid, //Style of the border
+            width: 4, //width of the border
+          ),
         ),
       ),
     );
