@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:zengage_learning_platform/constants/app_colors.dart';
+import 'package:zengage_learning_platform/extensions/hover_extensions.dart';
 
 // Nav links for routing to different pages
 class NavLink extends StatelessWidget {
@@ -7,26 +8,30 @@ class NavLink extends StatelessWidget {
   final String navigateToPage;
   final Color textColor;
   final bool isBold;
+  final double fontSize;
 
   NavLink(
       {this.navText,
       this.navigateToPage,
       this.textColor = textGreyColor,
-      this.isBold = false});
+      this.isBold = false,
+      this.fontSize = 18});
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      child: MaterialButton(
-        child: Text(
-          navText,
-          style: TextStyle(
-              color: textColor,
-              fontSize: 18.0,
-              fontWeight: isBold ? FontWeight.bold : FontWeight.normal),
+    return HandCursor(
+      child: Container(
+        child: MaterialButton(
+          child: Text(
+            navText,
+            style: TextStyle(
+                color: textColor,
+                fontSize: fontSize,
+                fontWeight: isBold ? FontWeight.bold : FontWeight.normal),
+          ),
+          onPressed: () => Navigator.popAndPushNamed(context,
+              navigateToPage), // TODO: implement the route it should navigate to
         ),
-        onPressed: () => Navigator.popAndPushNamed(context,
-            navigateToPage), // TODO: implement the route it should navigate to
       ),
     );
   }
