@@ -2,6 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:zengage_learning_platform/constants/app_colors.dart';
 import 'package:zengage_learning_platform/constants/app_constants.dart';
+import 'package:zengage_learning_platform/extensions/hover_extensions.dart';
 import 'package:zengage_learning_platform/models/DropDownValueItems.dart';
 import 'package:zengage_learning_platform/routes/route_generator.dart';
 import 'package:zengage_learning_platform/screens/home/widgets/social_banner.dart';
@@ -12,7 +13,7 @@ import 'package:zengage_learning_platform/widgets/upcoming_course_widgets/Filter
 
 class NavBar extends StatefulWidget implements PreferredSizeWidget {
   NavBar({Key key})
-      : preferredSize = Size.fromHeight(kToolbarHeight * 2),
+      : preferredSize = Size.fromHeight(kToolbarHeight * 2.7),
         super(key: key);
 
   @override
@@ -38,32 +39,34 @@ class _NavBarState extends State<NavBar> {
             onTap: () {
               Navigator.pushNamed(context, RouteGenerator.HOME_ROUTE);
             },
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.start,
-              crossAxisAlignment: CrossAxisAlignment.center,
-              children: <Widget>[
-                Container(
-                  height: 80.0,
-                  width: 80.0,
-                  alignment: Alignment.topCenter,
+            child: HandCursor(
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.start,
+                crossAxisAlignment: CrossAxisAlignment.center,
+                children: <Widget>[
+                  Container(
+                    height: 80.0,
+                    width: 80.0,
+                    alignment: Alignment.topCenter,
 //            padding: EdgeInsets.only(top: 100.0),
-                  decoration: BoxDecoration(
-                      image: DecorationImage(
-                          image: AssetImage(
-                              "assets/images/logos/company-logo.png"),
-                          fit: BoxFit.fill)),
-                ),
-                Container(
-                  margin: EdgeInsets.only(left: 16.0),
-                  child: Text(
-                    appName,
-                    style: TextStyle(
-                        color: blueThemeColor,
-                        fontSize: 26.0,
-                        fontWeight: FontWeight.bold),
+                    decoration: BoxDecoration(
+                        image: DecorationImage(
+                            image: AssetImage(
+                                "assets/images/logos/company-logo.png"),
+                            fit: BoxFit.fill)),
                   ),
-                )
-              ],
+                  Container(
+                    margin: EdgeInsets.only(left: 16.0),
+                    child: Text(
+                      appName,
+                      style: TextStyle(
+                          color: blueThemeColor,
+                          fontSize: 26.0,
+                          fontWeight: FontWeight.bold),
+                    ),
+                  )
+                ],
+              ),
             ),
           ),
           Expanded(
@@ -114,6 +117,11 @@ class _NavBarState extends State<NavBar> {
             children: <Widget>[
               SignInSignUp(),
 //              TODO: Add View Courses between the in Blue bg and White Text
+              NavLink(
+                navText: "View Courses",
+                isBold: true,
+                navigateToPage: RouteGenerator.UPCOMING_COURSE_ROUTE,
+              ),
               SocialBanner(),
             ],
           )
